refactor: move frozen structures to util submodule
